+++ src/pages/RestaurantsPage/RegionsContainer/RegionsContainer.jsx
@@ -1,6 +1,7 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import { selectRegion, loadRestaurants } from '@/store/actions';
+import { selectRegion } from '@/store/actions';
+import { loadRestaurants } from '@/store/async-actions';
 
 import { get } from '@/utils';
