+++ src/store/actions.js
@@ -1,11 +1,3 @@
-import {
-  fetchRegions,
-  fetchCategories,
-  fetchRestaurants,
-  fetchRestaurant,
-  postLogin,
-} from '@/services/api';
-
 export function setRegions(regions) {
   return {
     type: 'setRegions',
@@ -59,55 +51,5 @@ export function setAccessToken(accessToken) {
   return {
     type: 'setAccessToken',
     payload: { accessToken },
-  };
-}
-
-export function loadInitialData() {
-  return async (dispatch) => {
-    const regions = await fetchRegions();
-    dispatch(setRegions(regions));
-
-    const categories = await fetchCategories();
-    dispatch(setCategories(categories));
-  };
-}
-
-export function loadRestaurants() {
-  return async (dispatch, getState) => {
-    const { selectedRegion: region, selectedCategory: category } = getState();
-
-    if (!region || !category) {
-      return;
-    }
-
-    const restaurants = await fetchRestaurants({
-      regionName: region.name,
-      categoryId: category.id,
-    });
-    dispatch(setRestaurants(restaurants));
-  };
-}
-
-export function loadRestaurant({ restaurantId }) {
-  return async (dispatch) => {
-    dispatch(setRestaurant(null));
-
-    const restaurant = await fetchRestaurant({ restaurantId });
-
-    dispatch(setRestaurant(restaurant));
-  };
-}
-
-export function requestLogin() {
-  return async (dispatch, getState) => {
-    const { loginFields: { email, password } } = getState();
-
-    if (!email || !password) {
-      return;
-    }
-
-    const accessToken = await postLogin({ email, password });
-
-    dispatch(setAccessToken(accessToken));
   };
 }
